feat: folders refactoring
